Move maybe_evaluate to grpo_utils; dedupe calculate_token_counts Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
